telegram-bot/requirements.txt
python-telegram-bot==21.6
firebase-admin==6.5.0
